Make showing the version changes when listing available updates the default behaviour
Before that commit, the default behaviour was to not show the version changes when listing the available repo/AUR packages updates. Showing them was an option.

This commit aims to revert that behaviour (showing the version changes by default and make hiding them an option).
Indeed, version changes is an important and useful information which makes more sense being shown by default.

The documentation has been adapted accordingly regarding the required command to hide version changes for people that prefer not showing them.