Add an explicit message that the script is looking for updates
On a poor network connection and/or with a long list of pending updates, the script might take some time to look for and generate the list of packages available to update.

Without an explicit/clear indicator that the script is actually looking for updates, one could think that it is 'stuck'.
This commit introduces an explicit message that the script is currently looking for updates to avoid people thinking it hanged up.