Add an option in the configuration file to explicitely set which elevation command to use
This commit introduces the 'PrivilegeElevationCommand' option to the arch-update.conf configuration file which allows to explicitely set the command to be used to elevate privileges (sudo, doas, run0), so people can choose between if they have multiple of them installed.

If this option isn't set, the default behavior is applied (first look for `sudo`, then `doas`, then `run0`).

Closes https://github.com/Antiz96/arch-update/issues/173