chore: Rename "state" functions & file to "(tray) icon"
The "current_state" file and the related functions are in fact only used to determine which icon the systray applet should display (depending on the current state of the system).
Recording the current state of the system (up to date/updates available) is not necessary for any other stages (not currently at least), so it makes sense adopting a more suited & precise name.