chore: Make icon installations dynamic in the Makefile
Some new icons color set for the systray applet will be added soon and we are planning to make additional users contributed colors as easy as possible to implement.
Adoption a more "dynamic" way to install icons in the Makefile will avoid having to modify it each time a new icons color set is added.