feat(systray): Add dropdown sub-menus for earch sources of available updates (#348)
Divide the systray applet dropdown menu into sub-menus for each source of available updates (All, Package, AUR, Flatpak), making it easier to distinguish the number and the list of packages available for updates for each of them.

Sub-menus for sources with 0 pending update are not shown.
The "All" sub-menu is only shown when relevant (meaning when at least two sources have pending updates).
The "NoVersion" configuration file option is honored in sub-menus.

Closes https://github.com/Antiz96/arch-update/issues/347