fix(systray): Drop sub-menus for update sources in favor of "top level" dropdown menus
Some panels open dropdown menus in an arbitrary direction depending on different parameters such as the position of the applet in the tray area, the content length, the screen size, etc.. With dropdown sub-menus, this can result in overlaps on the main menu, preventing the main menu from being seen / used.

To prevent that, the dropdown sub-menus for each update sources are dropped in favor of "top level" dropdown menus (which cannot overlap with the main menu), right under the "update count" menu entry.

Fixes https://github.com/Antiz96/arch-update/issues/354