fix(systray): Drop sub-menus for update sources in favor of "top level" dropdown menus (#355)
Some panels open dropdown menus in an arbitrary direction depending on different parameters such as the position of the applet in the tray area, the content length, the screen size, etc.. With dropdown sub-menus, this can result in overlaps on the main menu, preventing the main menu from being seen / used (see screenshot below).

To prevent that, the dropdown sub-menus for each update sources are dropped in favor of "top level" dropdown menus (which cannot overlap with the main menu), right under the "update count" menu entry (see screenshot below).

They work the same way as the previous sub-menus:

- Dropdown menus for sources with 0 pending update are not shown.
- The "All" dropdown menu is only shown when relevant (meaning when at least two sources have pending updates).
- The "NoVersion" configuration file option is honored in dropdown menus.

Aside from the potential overlapping issue, this is arguably a better design as it requires less actions / mouse movements to get the number / list of pending updates.

Fixes https://github.com/Antiz96/arch-update/issues/354