feat: Make the blue systray icon the default
The "light" and "dark" systray icon are barely visible when using a theme of the same tone.
The "blue" one has the advantage of being visible on most theme variants and therefore might be a better default.

Users can still switch to their prefered icon variant via the `TrayIconStyle` option in the `arch-update.conf` configuration file.