fix: Skip kernel upgrade check if the kernel-modules-hook package is installed (#397)
The `[kernel-modules-hook](https://archlinux.org/packages/extra/any/kernel-modules-hook/)` package allows to keep the previous kernel modules loaded after a kernel upgrade. This avoids requiring a reboot after a kernel upgrade to load related modules. 

As such, running the "kernel upgrade check" function for users of the `kernel-modules-hook` package is not so relevant (as a reboot is not *strictly* required for them) and it results in a confusing output (as it says that no pending kernel upgrade is detected while there actually might be one).

Fixes https://github.com/Antiz96/arch-update/issues/396