fix(notification): Switch to `systemd-run` to detach desktop notifications processes
Switch from `disown` to `systemd-run` for detaching desktop notifications processes.
This is to fix the issue where the automatic desktop notifications sent by the systemd timer are not visible to users (due to the fact that `disown` doesn't expand environment variables to the subshell).

Fixes https://github.com/Antiz96/arch-update/issues/400