fix(news): Make URL path parsing more robust
The current logic for parsing news URL assumed that the URL path would always match the title of the news entry, which turned out to not necessarily be the case, resulting in eventual wrongly parsed URL paths.

This commit makes the logic to parse URL paths more robust by relying on the actual path as defined as `href` in the https://archlinux.org/news source code (and, as a bonus, it also allows to *finally* get rid of that `sed` monstrosity).