Add tray to user's XDG autostart when starting the systemd service
This allows `CachyOS Hello` to set the tray to run both via the systemd service and XDG autostart, covering cases for most / all DEs depending on what they support (for instance, XFCE doesn't play well with the systemd service and therefore needs the tray to be started via XDG autostart instead).

There's no concern having the tray started both as a systemd service and via XDG autostart as it is smart enough to not start if there's already a process running.