fix(notification): Fix "Run Arch-Update" click action
The "Run Arch-Update" click action in the desktop notifications was broken due to the related "lockfile" variable not being passed correctly to the `systemd-run` environment the notification library is executed in.

After a second thought, there's no need for those "lockfiles" to be put in variables in the first place. They can just be called explicitly where needed.