Rebase against v3.16.0 upstream release (#7)
* Rebase against upstream tree (v3.14.0) (#2)

https://github.com/Antiz96/arch-update/releases/tag/v3.14.0

* Add tray to user's XDG autostart when starting the systemd service

This allows `CachyOS Hello` to set the tray to run both via the systemd service and XDG autostart, covering cases for most / all DEs depending on what they support (for instance, XFCE doesn't play well with the systemd service and therefore needs the tray to be started via XDG autostart instead).

There's no concern having the tray started both as a systemd service and via XDG autostart as it is smart enough to not start if there's already a process running.

* Update tray tooltip to match CachyOS branding (#3)

* Add slight delays before starting Tray to avoid eventual race condition

* Add slight delays before starting Tray to avoid eventual race condition

* Revert workaround to prevent eventual race condition

This is now properly fixed upstream

* Update README.md for CachyOS fork (#4)

* doc: Update README.md for CachyOS fork

* Add a link to Arch-Update

---------

Co-authored-by: Robin Candau <robincandau@protonmail.com>

* Update pt_BR translation to match CachyOS branding

* Update es translation to match CachyOS branding

* Update po files

---------

Co-authored-by: Niko Diamadis <cyb3rko@pm.me>