fix: Fix orphan packages removal with `sudo-rs` & `run0` (#456)
`sudo-rs` and `run0` do not forward stdin when piped through, causing
interactive commands relying on piped stdin (such as the one currently
used to remove orphan packages) to exit prematurely / crash.

We therefore now extract the list of orphan packages into an array and
rely on it to pass the list of packages to remove to pacman (rather than
relying on a piped stdin).

Fixes https://github.com/Antiz96/arch-update/issues/455